# Conflicts:
#	buildspec.yml
#	vite.config.js

+++ src/components/service/environment/personalInfo/PersonalInfoCategory.svelte
@@ -1,7 +1,9 @@
 <script>
-    import {personalInfoCategoryData} from "../../../../lib/store.js";
+    import {personalInfoCategoryData, userInfoData} from "../../../../lib/store.js";
     import {onDestroy, onMount} from "svelte";
     export let personalInfoCategoryService;
+
+    $: isModifiable = ['ROLE_MASTER', 'ROLE_ADMIN'].includes($userInfoData.role);
 
     // 카테고리를 선택했을 때 해당 카테고리에 맞는 항목들을 보여주기 위함
     const catBoxControl = (e) => {
@@ -21,11 +23,13 @@
     });
 </script>
 
-<div class="prPart1_box">
+<div class="prPart1_box" style="width: 59%">
     <div class="prptitle">
         <h2>항목 분류</h2>
         <!-- svelte-ignore a11y-click-events-have-key-events -->
-        <div class="myAddBtn" on:click={personalInfoCategoryService.createItemPop.show}>나만의 항목 추가</div>
+        {#if isModifiable}
+            <div class="myAddBtn" on:click={personalInfoCategoryService.createItemPop.show}>나만의 항목 추가</div>
+        {/if}
     </div>
     <div class="categorydivision_box">
         <div>
@@ -79,14 +83,14 @@
 
                         {#if $personalInfoCategoryData.addItemList.length}
                             <div class="cateS_checkInner">
-                                {#each $personalInfoCategoryData.addItemList as {ciName, categoryName}, i}
+                                {#each $personalInfoCategoryData.addItemList as {ciName, categoryName, ciSecurity}, i}
                                     <div class="cateS_check">
                                         <input type="checkbox" name="itemCheck" id="cates_0{i}" value={ciName}
                                                bind:group={$personalInfoCategoryData.checkedItemNameList}
                                                on:change={personalInfoCategoryService.handleCheckedItemChange}>
                                         <label for="cates_0{i}">
                                             <em></em>
-                                            <p class="check">{ciName}</p>
+                                            <p class="check">{#if ciSecurity}<span class="lockicon"></span>{/if}{ciName}</p>
                                         </label>
                                         <button on:click={()=>{personalInfoCategoryService.editItemPop.show(i)}} class="cateReviseIcon"></button>
                                     </div>
@@ -96,18 +100,19 @@
 
                         {#each $personalInfoCategoryData.basicCategoryList as {cdName, categoryItemListDtoList}, i}
                             <div class="cateS_checkInner" style="display:none;">
-                                {#each categoryItemListDtoList as {ciName, categoryName, textColor}, j}
+                                {#each categoryItemListDtoList as {ciName, categoryName, textColor, ciSecurity}, j}
                                     <div class="cateS_check">
                                         <input type="checkbox" name="itemCheck" id="cates_{i + 1}{j}" value={ciName}
                                                bind:group={$personalInfoCategoryData.checkedItemNameList}
                                                on:change={personalInfoCategoryService.handleCheckedItemChange}>
                                         <label for="cates_{i + 1}{j}">
                                             <em></em>
-                                            <!-- 암호화 여부의 따라 해당 아이콘 넣어야 될지도? <div class="lockicon"></div>-->
-                                            <p class="check">{ciName}</p>
+                                            <p class="check">{#if ciSecurity}<span class="lockicon"></span>{/if}{ciName}</p>
                                         </label>
+                                        {#if categoryName}
                                         <!-- 클래스 받은값으로 색변경 -->
-                                        <span class="subElement {textColor}">{categoryName}</span>
+                                            <span class="subElement {textColor}">{categoryName}</span>
+                                        {/if}
                                     </div>
                                 {/each}
                             </div>
@@ -117,20 +122,22 @@
             </div>
         </div>
 
-        <div class="sel_cateListBox">
-            <div class="sel_cateList">
-                {#each $personalInfoCategoryData.checkedItemObjList as {ciName}, i}
-                    <div class="sel_cate" id="choseCategory{i}">
-                        {ciName}
-                        <button on:click={() => personalInfoCategoryService.removeCheckedItem(ciName)}>X</button>
-                    </div>
-                {/each}
+        {#if isModifiable}
+            <div class="sel_cateListBox">
+                <div class="sel_cateList">
+                    {#each $personalInfoCategoryData.checkedItemObjList as {ciName}, i}
+                        <div class="sel_cate" id="choseCategory{i}">
+                            {ciName}
+                            <button on:click={() => personalInfoCategoryService.removeCheckedItem(ciName)}>X</button>
+                        </div>
+                    {/each}
+                </div>
+                <div class="sel_cateBtnBox">
+                    <button class="cateResetBtn" on:click={personalInfoCategoryService.resetCheckedItemState}>초기화</button>
+                    <button class="cateAddBtn" on:click={personalInfoCategoryService.insertItemPop.handleAddItemBtnClick}>오른쪽에 추가</button>
+                </div>
             </div>
-            <div class="sel_cateBtnBox">
-                <button class="cateResetBtn" on:click={personalInfoCategoryService.resetCheckedItemState}>초기화</button>
-                <button class="cateAddBtn" on:click={personalInfoCategoryService.insertItemPop.handleAddItemBtnClick}>오른쪽에 추가</button>
-            </div>
-        </div>
+        {/if}
 <!--                    <div class="cateAddInfoBox">-->
 <!--                        <div class="cateAddInfo" id="commerce_pop">전자상거래법 적용 대상이세요?</div><br>-->
 <!--                        <div class="cateAddInfo" id="create_item_pop">원하시는 개인정보 항목이 없나요?</div>-->
